+++ crates/kira/src/manager/backend/resources/modulators.rs
@@ -19,13 +19,10 @@ pub(crate) struct Modulators {
 }
 
 impl Modulators {
-	pub fn new(
-		capacity: usize,
-		unused_modulator_producer: HeapProducer<BufferedModulator>,
-	) -> Self {
+	pub fn new(capacity: u16, unused_modulator_producer: HeapProducer<BufferedModulator>) -> Self {
 		Self {
 			modulators: Arena::new(capacity),
-			modulator_ids: Vec::with_capacity(capacity),
+			modulator_ids: Vec::with_capacity(capacity as usize),
 			unused_modulator_producer,
 			dummy_modulator: BufferedModulator::new(Box::new(DummyModulator)),
 		}
